+++ test/message_test.dart
@@ -18,9 +18,9 @@ void main() {
       var server = Client();
       await server.connect('localhost');
       var service = server.sub('service');
-      unawaited(service.stream.first.then((m) {
-        m.respond(Uint8List.fromList('respond'.codeUnits));
-      }));
+      service.stream.listen((m) {
+        m.respondString('respond');
+      });
 
       var requester = Client();
       await requester.connect('localhost');